refactor: move helpers to the root folder
