+++ src/__tests__/mocks/link.ts
